Fix NodePattern for fields without a name
`field` can be called without a position name argument which causes
errors right now. This also happens when FieldExtension's are used;
methods are called on `field` instead of passing arguments.

The simplest fix was to ensure the node pattern matching for field
definitions required a symbol positional argument (the field name).