Check a variety of base classes in the NotAuthorizedNodeType check
When a class inherits from another the syntax Ruby uses isn't precise
and means that when looking at a single file we need to check a two
different names for the superclass.  One assumes the given superclass is
absolute, and the second preprepends any parent modules/classes of the
class we're looking at.

This change builds a list of up to two classes to use when checking the
SafeBaseClasses.